Considero sleeps de más de 8 horas

+++ alimCalle.h
@@ -11,7 +11,7 @@
 void initServo(void);
 void mueveServoAncho(uint16_t ancho);
 void mueveServoPos(uint8_t porcPosicion);
-uint8_t sleep_for_x_sec(uint16_t nb_sec);
+uint8_t sleep_for_x_sec(uint32_t nb_sec);
 uint8_t stop(uint16_t nb_sec);
 uint8_t standby0(uint16_t nb_sec);
 void ports_set_lowpower(void);
